Add plugins API
This adds to tmux the ability to load "native" plugins, with full
access to tmux's internals.

This is a very powerful capability, but also inherently unstable:
even otherwise nonbreaking changes to tmux's internals can break
this kind of plugin. Still, this adds a layer of flexibility and
customizability that you just can't get otherwise, at least not
without forking and patching tmux.

At it's core, that's what this plugin system is meant to replace,
not things that could otherwise be done with tmux's very flexible
scripting interface. In fact, the main API surface exposed to plugins
is the ability to build on tmux's formatting engine, to extend what
is offered via tmux's scripting interface. It is expected that most
plugins will not implement their core functionality directly, but
rather extend the capability of what can be done via tmux's
configuration.